backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
langchain==0.1.13
langchain-core==0.1.33
langchain-experimental==0.0.54
perplexity-ai==0.0.13
firebase-admin==6.4.0
python-socketio==5.11.1
websockets==12.0
pytest==7.4.3
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.3
python-jose==3.3.0
aiohttp==3.9.3 